@@ -3,7 +3,7 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.black]
-target-version = ["py37", "py38", "py39", "py310", "py311"]
+target-version = ["py39", "py310", "py311"]
 
 [tool.flakeheaven]
 max-complexity = 10
@@ -41,7 +41,7 @@ readme = "README.md"
 repository = "https://github.com/christophgietl/pisco"
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.9"
 click = "^8.1"
 Pillow = "^9.4"
 python-json-logger = "^2.0"
@@ -51,7 +51,7 @@ xdg = "^5.1"
 
 [tool.poetry.dev-dependencies]
 mypy = "^1.0"
-pre-commit = "^2.21"
+pre-commit = "^3.0"
 pytest = "^7.2"
 pytest-cov = "^4.0"
 types-Pillow = "^9.4"
