@@ -1,4 +1,4 @@
-from pathlib import Path
+import pathlib
 
 
 def test_pytest() -> None:
@@ -12,7 +12,9 @@ class TestBacklight:
         backlight = Backlight(backlight_directory=None)
         backlight.activate()
 
-    def test_activate_sets_brightness_to_max_brightness(self, tmp_path: Path) -> None:
+    def test_activate_sets_brightness_to_max_brightness(
+        self, tmp_path: pathlib.Path
+    ) -> None:
         brightness_file_path = tmp_path / "brightness"
         brightness_file_path.write_text("19")
         max_brightness_file_path = tmp_path / "max_brightness"
@@ -30,7 +32,7 @@ class TestBacklight:
         backlight = Backlight(backlight_directory=None)
         backlight.deactivate()
 
-    def test_deactivate_sets_brightness_to_0(self, tmp_path: Path) -> None:
+    def test_deactivate_sets_brightness_to_0(self, tmp_path: pathlib.Path) -> None:
         brightness_file_path = tmp_path / "brightness"
         brightness_file_path.write_text("19")
 
